Create controleer for musican scills model
