+++ config/routes.rb
@@ -5,5 +5,6 @@ Rails.application.routes.draw do
     resources :places
     resources :users
     resources :organizations
+    resources :musician_skill
   end
 end
