+++ db/schema.rb
@@ -11,8 +11,6 @@
 # It's strongly recommended that you check this file into your version control system.
 
 ActiveRecord::Schema.define(version: 2019_08_09_111639) do
-
-  # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
   create_table "events", force: :cascade do |t|
@@ -34,8 +32,8 @@ ActiveRecord::Schema.define(version: 2019_08_09_111639) do
   end
 
   create_table "images", force: :cascade do |t|
-    t.integer "imageable_id", null: false
-    t.string "imageable_type", null: false
+    t.bigint "imageable_id"
+    t.string "imageable_type"
     t.string "url", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
@@ -48,6 +46,7 @@ ActiveRecord::Schema.define(version: 2019_08_09_111639) do
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.integer "status", default: 0, null: false
+
     t.index ["musician_skill_id"], name: "index_musician_skill_users_on_musician_skill_id"
     t.index ["user_id"], name: "index_musician_skill_users_on_user_id"
   end
