--- spec/controllers/musican_scills_controller_controller_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe MusicanScillsControllerController, type: :controller do
-
-end
